ref/appendices.texi: Fixed reference to Top node of r4rs.texi. Removed unnumbered "Indices" node, it's not too important, and makeinfo complains because it doesn't appear as an item in the top-level menu.
ref/intro.texi (Intro): Sketch the outline of the manual in the intro node.
(Guile Scripts): Break this into two subnodes and add a menu, so we
can have a menu pointing to "Invoking Guile", and not have a
one-element menu.

ref/scm.texi: Remove chapter "Linking Guile with your code"; it's now
in the Preliminaries part.