* session.scm: Update references to `proc-doc' to be `proc-documentation'
* doc.scm: Cleaned up a great deal.  Put variables at the top of
the file, eliminated `object-documentation' that was broken
(referencing Scwm), drop `help' as session.scm has a better
supported version of that procedure. Rename `proc-doc' to
`proc-documentation' -- `procedure-documentation' is a primitive
getter function, so I use the shorter name for this more useful
function.  (Alternatively, we could rename the primitive
getter...)