* session.scm (apropos): Do not use `builtin-bindings', always use the module obarray. (apropos-fold): Likewise.
* optargs.scm (bound?): Removed.  We should not play games with
the magical undefined value.
(let-o-k-template): Use `#f' instead of the undefined value as
the default default for bindings.

* boot-9.scm (module-make-local-var!): Do not pass name hint to
make-undefined-variable.
(root-module-closure): Removed.
(make-root-module): Set the obarray of the module to the
`pre-modules-obarray'.  Do not use a lazy binder.
(scm-module-closure): Removed.
(make-root-module): Set the obarray of the module to the
`pre-modules-obarray'.  Do not use a lazy binder.  Set the
eval-closure to a `standard-interface-eval-closure'.
(module-define!): Do not pass name hint to make-variable.
(make-modules-in, beautify-user-module, resolve-module): Moved
towards the beginning of boot-9.scm, across the call to
set-current-module that boots the module system.  These
definitions need to be visible at the time of the first
`set-current-module' call.
(try-module-autoload): Define a `#f' before the call to
set-current-module.  It is redefined later.

* debug.scm: Use `module-set!' instead of `variable-set!' to set
insert `debug-options' into the-root-module.
* format.scm: Likewise, for `format'.